[Hardware][Intel GPU] Add Intel GPU(XPU) inference backend (#3814)

+++ vllm/model_executor/layers/vocab_parallel_embedding.py
@@ -307,7 +307,7 @@ class VocabParallelEmbedding(torch.nn.Module):
         else:
             masked_input = input_
             # Get the embeddings.
-        output_parallel = F.embedding(masked_input, self.weight)
+        output_parallel = F.embedding(masked_input.long(), self.weight)
         # Mask the output embedding.
         if self.tp_size > 1:
             output_parallel.masked_fill_(input_mask.unsqueeze(1), 0)
